% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztb.R
\name{ztb}
\alias{ztb}
\title{ZTB: A really fast HTML table viewer}
\usage{
ztb(tb_, use_viewer_ = F)
}
\arguments{
\item{tb_}{A \code{data.frame} or something that can be coerced into a \code{tibble} using \code{as_tibble}.}
}
\description{
This is for displaying your data frames or matrices on your browser. It is
particularly meant for large tables and speed is the first priority.
}

